extracting data calcul and adding activities count on résumé des echanges
