+++ src/components/admin/dashboard-statistics/charts/HorizontalChart.tsx
@@ -2,33 +2,34 @@ import React from 'react';
 
 import { BarChart } from '@mui/x-charts/BarChart';
 
-import { mockClassroomsStats } from '../mocks/mocks';
+// import { mockClassroomsStats } from '../mocks/mocks';
 import styles from '../styles/charts.module.css';
+import { sumData } from '../utils/sumData';
 
-interface SumData {
-  country: string;
-  total: number;
-}
+// interface SumData {
+//   country: string;
+//   total: number;
+// }
 
-const sumData: { [country: string]: SumData } = {};
-mockClassroomsStats.forEach((country) => {
-  const { classroomCountryCode, commentsCount, activities } = country;
+// const sumData: { [country: string]: SumData } = {};
+// mockClassroomsStats.forEach((country) => {
+//   const { classroomCountryCode, commentsCount, activities } = country;
 
-  if (!sumData[classroomCountryCode]) {
-    sumData[classroomCountryCode] = {
-      country: classroomCountryCode,
-      total: 0,
-    };
-  }
+//   if (!sumData[classroomCountryCode]) {
+//     sumData[classroomCountryCode] = {
+//       country: classroomCountryCode,
+//       total: 0,
+//     };
+//   }
 
-  sumData[classroomCountryCode].total += commentsCount;
+//   sumData[classroomCountryCode].total += commentsCount;
 
-  activities.forEach((activity) => {
-    sumData[classroomCountryCode].total += activity.count;
-  });
-});
+//   activities.forEach((activity) => {
+//     sumData[classroomCountryCode].total += activity.count;
+//   });
+// });
 
-const dataset: { country: string; total: number }[] = Object.values(sumData);
+// const dataset: { country: string; total: number }[] = Object.values(sumData);
 
 const chartSetting = {
   width: 500,
@@ -41,7 +42,7 @@ export default function HorizontalBars() {
   return (
     <div className={styles.horizontalBars}>
       <BarChart
-        dataset={dataset}
+        dataset={sumData}
         yAxis={[{ scaleType: 'band', dataKey: 'country' }]}
         series={[
           {
